api/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
cymem==2.0.8
deep-translator==1.11.4
Django==4.2.6
en-core-web-md @ https://huggingface.co/spacy/en_core_web_md/resolve/main/en_core_web_md-any-py3-none-any.whl#sha256=d37c3d60738be409a04f83a51421075b94fc02efad99bb8e62ec9018b1f99cb0
googlemaps==4.10.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
mysql-connector-python==8.1.0
numpy==1.26.0
packaging==23.2
pathy==0.10.2
preshed==3.0.9
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
requests==2.31.0
schedule==1.2.1
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.2

